feat: move enum to proper module instead of utils module
