+++ src/earning.rs
@@ -9,8 +9,7 @@ use serde::Deserialize;
 use crate::{
     api::ApiClient,
     deserialize::{from_none_str, from_str},
-    error::{Error, Result},
-    utils::detect_common_helper_error,
+    error::{detect_common_helper_error, Error, Result},
 };
 
 /// Struct to store information of annual earning
